# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An example of gin",
        "title": "Golang Gin API",
        "termsOfService": "运行地址：http://localhost/swagger/index.html",
        "contact": {},
        "license": {
            "name": "MIT //localhost:80"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/api/musics": {
            "get": {
                "description": "音乐列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "音乐"
                ],
                "summary": "获取音乐列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pageNum",
                        "name": "pageNum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Music"
                        }
                    },
                    "400": {
                        "description": "{ \"code\": 400, \"message\": \"请求失败\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "创建音乐",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "音乐"
                ],
                "summary": "创建音乐列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "year",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "author",
                        "name": "actor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "desc",
                        "name": "desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Music"
                        }
                    },
                    "400": {
                        "description": "{ \"code\": 400, \"message\": \"请求失败\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/musics/{id}": {
            "put": {
                "description": "音乐列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "音乐"
                ],
                "summary": "更新音乐列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Music"
                        }
                    },
                    "400": {
                        "description": "{ \"code\": 400, \"message\": \"id必传\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "音乐列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "音乐"
                ],
                "summary": "删除音乐列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Music"
                        }
                    },
                    "400": {
                        "description": "{ \"code\": 400, \"message\": \"id必传\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Music": {
            "type": "object",
            "properties": {
                "actor": {
                    "type": "string",
                    "example": "作曲"
                },
                "desc": {
                    "type": "string",
                    "example": "描述"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "example": "歌曲名称"
                },
                "year": {
                    "type": "string",
                    "example": "年份"
                }
            }
        },
...
    }
}